Backend

+++ src/main/java/it/unisa/studenti/nc8/gametalk/business/service/post/CommentServiceImpl.java
@@ -1,15 +1,24 @@
 package it.unisa.studenti.nc8.gametalk.business.service.post;
 
-import it.unisa.studenti.nc8.gametalk.business.validators.CommentValidator;
-import it.unisa.studenti.nc8.gametalk.business.validators.Validator;
-import it.unisa.studenti.nc8.gametalk.storage.exceptions.DAOException;
 import it.unisa.studenti.nc8.gametalk.business.exceptions.ServiceException;
+import it.unisa.studenti.nc8.gametalk.business.model.post.comment.Comment;
+import it.unisa.studenti.nc8.gametalk.business.model.post.thread.Thread;
+import it.unisa.studenti.nc8.gametalk.business.model.user.User;
+import it.unisa.studenti.nc8.gametalk.business.validators.Validator;
+import it.unisa.studenti.nc8.gametalk.business.validators.post.comment.CommentValidator;
 import it.unisa.studenti.nc8.gametalk.storage.dao.post.comment.CommentDAO;
 import it.unisa.studenti.nc8.gametalk.storage.dao.post.comment.CommentDAOImpl;
+import it.unisa.studenti.nc8.gametalk.storage.dao.post.thread.ThreadDAO;
+import it.unisa.studenti.nc8.gametalk.storage.dao.post.thread.ThreadDAOImpl;
+import it.unisa.studenti.nc8.gametalk.storage.dao.user.UserDAO;
+import it.unisa.studenti.nc8.gametalk.storage.dao.user.UserDAOImpl;
+import it.unisa.studenti.nc8.gametalk.storage.exceptions.DAOException;
 import it.unisa.studenti.nc8.gametalk.storage.persistence.Database;
 import it.unisa.studenti.nc8.gametalk.storage.persistence.mappers.comment.CommentMapper;
-import it.unisa.studenti.nc8.gametalk.business.model.post.comment.Comment;
+import it.unisa.studenti.nc8.gametalk.storage.persistence.mappers.thread.ThreadMapper;
+import it.unisa.studenti.nc8.gametalk.storage.persistence.mappers.user.UserMapper;
 
+import java.sql.SQLException;
 import java.time.LocalDate;
 import java.util.List;
 
@@ -19,10 +28,27 @@ import java.util.List;
 public class CommentServiceImpl implements CommentService {
 
     /**
+     * Oggetto {@link Database}, contiene informazioni sul database
+     * con cui interagire.
+     */
+    private final Database db;
+    /**
      * Il DAO utilizzato per effettuare operazioni CRUD
      * su oggetti {@link Comment}.
      */
     private final CommentDAO commentDAO;
+
+    /**
+     * Il DAO utilizzato per effettuare operazioni CRUD
+     * su oggetti {@link User}.
+     */
+    private final UserDAO userDAO;
+
+    /**
+     * Il DAO utilizzato per effettuare operazioni CRUD
+     * su oggetti {@link Thread}.
+     */
+    private final ThreadDAO threadDAO;
 
     /**
      * Il validator che valida i dati contenuti in
@@ -36,7 +62,10 @@ public class CommentServiceImpl implements CommentService {
      * @param db il database utilizzato per la persistenza dei dati.
      */
     public CommentServiceImpl(final Database db) {
+        this.db = db;
         this.commentDAO = new CommentDAOImpl(db, new CommentMapper());
+        this.userDAO = new UserDAOImpl(db, new UserMapper());
+        this.threadDAO = new ThreadDAOImpl(db, new ThreadMapper());
         this.commentValidator = new CommentValidator();
     }
 
@@ -44,40 +73,57 @@ public class CommentServiceImpl implements CommentService {
      * Aggiunge un nuovo commento a un thread esistente e lo salva nel database.
      * Il commento viene validato prima di essere salvato.
      *
-     * @param id L'ID del nuovo commento.
      * @param threadId L'ID del thread a cui il commento appartiene.
-     * @param userId L'ID dell'utente che ha scritto il commento.
+     * @param username L'ID dell'utente che ha scritto il commento.
      * @param body Il corpo del commento.
-     * @param votes Il numero di voti iniziali del commento.
      *
-     * @throws ServiceException Se il commento non è valido o se si verifica
-     *                          un errore durante il salvataggio nel database.
+     * @throws ServiceException Se si verifica un errore durante il
+     * salvataggio nel database.
+     * @throws IllegalArgumentException Se il commento non è valido.
      */
-    public void addComment(final long id,
-                           final long threadId,
-                           final long userId,
-                           final String body,
-                           final int votes)
-            throws ServiceException {
+    public void addComment(
+            final long threadId,
+            final String username,
+            final String body
+    ) throws ServiceException, IllegalArgumentException {
 
         //Inizializzazione oggetto Comment
         Comment newComment = new Comment();
-        newComment.setId(id);
         newComment.setThreadId(threadId);
-        newComment.setUserId(userId);
+        newComment.setUsername(username);
         newComment.setBody(body);
-        newComment.setVotes(votes);
+        newComment.setVotes(0);
         newComment.setCreationDate(LocalDate.now());
 
         //Validazione commento
         if (!commentValidator.validate(newComment)) {
-            throw new ServiceException("Commento non valido");
+            throw new IllegalArgumentException("Commento non valido");
         }
 
         //Salvataggio comment
-        try {
+        try (db) {
+            db.connect();
+            db.beginTransaction();
+
+            //Verifica thread non archiviato
+            Thread thread = threadDAO.get(threadId);
+            if (thread == null) {
+                throw new ServiceException(
+                        "Nessun thread trovato con id " + threadId);
+            }
+
+            if (thread.isArchived()) {
+                throw new ServiceException("Thread archiviato");
+            }
+
             commentDAO.save(newComment);
-        } catch (DAOException e) {
+            db.commit();
+        } catch (SQLException | DAOException e) {
+            try {
+                db.rollback();
+            } catch (SQLException ex) {
+                throw new ServiceException(ex);
+            }
             throw new ServiceException(
                     "Errore durante il salvataggio del commento", e);
         }
@@ -92,14 +138,39 @@ public class CommentServiceImpl implements CommentService {
      */
     @Override
     public void deleteComment(final long id) throws ServiceException {
-        try {
             if (id <= 0) {
                 throw new IllegalArgumentException(
                         "Id deve essere maggiore di 0");
             }
-            commentDAO.delete(id);
-        } catch (DAOException e) {
-            throw new RuntimeException(e);
+            try (db) {
+                db.connect();
+                db.beginTransaction();
+                //Verifica thread non archiviato
+                Comment comment = commentDAO.get(id);
+                if (comment == null) {
+                    throw new ServiceException(
+                            "Comment non trovato con id " + id);
+                }
+
+                Thread thread = threadDAO.get(comment.getThreadId());
+                if (thread == null) {
+                    throw new ServiceException("Thread non trovato");
+                }
+
+                if (thread.isArchived()) {
+                    throw new ServiceException("Thread archiviato");
+                }
+
+                commentDAO.delete(id);
+                db.commit();
+        } catch (SQLException | DAOException e) {
+                try {
+                    db.rollback();
+                } catch (SQLException ex) {
+                    throw new ServiceException(ex);
+                }
+                throw new ServiceException(
+                    "Errore durante l'eliminazione del commento", e);
         }
     }
 
@@ -112,14 +183,17 @@ public class CommentServiceImpl implements CommentService {
      * @throws ServiceException se si è verificato un errore.
      */
     public Comment findCommentById(final long id) throws ServiceException {
-        try {
-            if (id <= 0) {
-                throw new IllegalArgumentException(
-                        "ID deve essere maggiore di 0");
-            }
+        if (id <= 0) {
+            throw new IllegalArgumentException(
+                    "ID deve essere maggiore di 0");
+        }
 
+        try (db) {
+
+            db.connect();
             return commentDAO.get(id);
-        } catch (DAOException e) {
+
+        } catch (SQLException | DAOException e) {
             throw new ServiceException(
                     "Errore recupero commento tramite ID.", e);
         }
@@ -129,6 +203,8 @@ public class CommentServiceImpl implements CommentService {
      * Recupera i commenti di un thread, con supporto per la paginazione.
      *
      * @param threadId Il ID del thread di cui recuperare i commenti.
+     * @param userName Il nome utente del richiedente <code>null</code>
+     *                 se non è loggato.
      * @param page     Il numero della pagina da recuperare.
      * @param pageSize Il numero di commenti per pagina.
      * @return Una lista di commenti del thread specificato.
@@ -137,15 +213,16 @@ public class CommentServiceImpl implements CommentService {
      * @throws ServiceException se si è verificato un errore.
      */
     @Override
-    public List<Comment> findCommentsFromThreadId(
+    public List<Comment> findCommentsByThreadId(
             final long threadId,
+            final String userName,
             final int page,
             final int pageSize
     ) throws ServiceException {
-
         //Sanificazione TODO da spostare
         int realPage = Math.max(page, 1);
 
+        String realUserName = (userName == null) ? "" : userName;
         //Verifica id valido
         if (threadId <= 0) {
             throw new IllegalArgumentException(
@@ -153,17 +230,136 @@ public class CommentServiceImpl implements CommentService {
         }
 
         //Recupero commenti dal thread
-        try {
-
+        try (db) {
+            db.connect();
             return commentDAO.getCommentsByThreadId(
                     threadId,
+                    realUserName,
                     realPage,
                     pageSize);
 
-        } catch (DAOException e) {
+        } catch (SQLException | DAOException e) {
             throw new ServiceException(
                     "Errore durante il recupero dei "
                             + "commenti appartenenti al thread " + threadId, e);
+        }
+    }
+
+    /**
+     * Restituisce il numero totale di commenti di un thread.
+     *
+     * @param threadId Il ID del thread di cui recuperare i commenti.
+     * @return il numero totale di commenti del thread.
+     * @throws IllegalArgumentException se il <code>threadId</code>
+     *                                  è minore o uguale a 0
+     * @throws ServiceException         se si è verificato un errore.
+     */
+    @Override
+    public long countCommentsByThreadId(
+            final long threadId
+    ) throws ServiceException {
+        // Verifica id valido
+        if (threadId <= 0) {
+            throw new IllegalArgumentException(
+                    "threadId deve essere maggiore di 0");
+        }
+
+        // Recupero commenti dal thread
+        try (db) {
+            db.connect();
+            return commentDAO.countCommentsByThreadId(threadId);
+        } catch (SQLException | DAOException e) {
+            throw new ServiceException(
+                    "Errore durante il conteggio dei "
+                            + "commenti appartenenti al thread " + threadId, e);
+        }
+    }
+
+    /**
+     * Permette a un utente di votare un commento, con la possibilità di
+     * rimuovere un voto esistente (impostando il voto a 0). In caso di
+     * voto invalido, o se il commento non esiste, viene sollevata un'eccezione.
+     *
+     * @param commentId ID del commento da votare.
+     * @param username Nome utente dell'utente che sta effettuando il voto.
+     * @param vote Valore del voto da assegnare al commento, deve essere:
+     *             <ul>
+     *             <li>-1: Downvote.</li>
+     *             <li>0: Rimozione del voto esistente (se presente).</li>
+     *             <li>1: Upvote.</li>
+     *             </ul>
+     *
+     * @throws ServiceException Se si verifica un errore durante l'elaborazione
+     * del voto, come:
+     * <ul>
+     * <li>Il commento con l'ID specificato non esiste.</li>
+     * <li>Errore durante l'aggiunta del voto.</li>
+     * </ul>
+     * @throws IllegalArgumentException Se il valore del voto non è valido
+     * (diverso da -1, 0, 1).
+     *
+     */
+    @Override
+    public void rateComment(
+            final long commentId,
+            final String username,
+            final int vote
+    ) throws ServiceException {
+
+        //Sanificazione
+        if (vote < -1 || vote > 1) {
+            throw new IllegalArgumentException("Voto non valido");
+        }
+
+        try (db) {
+
+            db.connect();
+            db.beginTransaction();
+
+            //Verifico l'esistenza del commento
+            Comment comment = commentDAO.get(commentId);
+            if (comment == null) {
+                throw new ServiceException(
+                        "Nessun commento trovato con id " + commentId);
+            }
+
+            //Verifico l'esistenza dell'utente
+            User user = userDAO.get(username);
+            if (user == null) {
+                throw new ServiceException(
+                        "Nessun utente trovato con username " + username);
+            }
+
+            //Verifico se il thread a cui appartiene il commento è archiviato
+            Thread thread = threadDAO.get(comment.getThreadId());
+            if (thread == null) {
+                throw new ServiceException("Thread non trovato");
+            }
+
+            if (thread.isArchived()) {
+                throw new ServiceException("Thread archiviato");
+            }
+
+            //Commento esiste, lo voto da parte dell utente
+            if (vote == 0) {
+                //Rimuovo il voto
+                commentDAO.removeVoteComment(commentId, username);
+            } else {
+                //Inserisco il voto
+                commentDAO.voteComment(commentId, username, vote);
+            }
+
+            db.commit();
+
+        } catch (SQLException | DAOException e) {
+            try {
+                db.rollback();
+            } catch (SQLException ex) {
+                throw new ServiceException(ex);
+            }
+            throw new ServiceException(
+                    "Errore durante l'aggiunta del voto " + vote
+                            + " al commento " + commentId, e);
         }
     }
 }
